+++ include/fastrtps/rtps/reader/StatefulReader.h
@@ -41,240 +41,247 @@ class RTPSMessageSenderInterface;
  */
 class StatefulReader : public RTPSReader
 {
-    public:
+public:
 
-        friend class RTPSParticipantImpl;
+    friend class RTPSParticipantImpl;
 
-        virtual ~StatefulReader();
+    virtual ~StatefulReader();
 
-    protected:
+protected:
 
-        StatefulReader(
-                RTPSParticipantImpl*,
-                const GUID_t& guid,
-                const ReaderAttributes& att,
-                ReaderHistory* hist,
-                ReaderListener* listen = nullptr);
-    public:
+    StatefulReader(
+            RTPSParticipantImpl*,
+            const GUID_t& guid,
+            const ReaderAttributes& att,
+            ReaderHistory* hist,
+            ReaderListener* listen = nullptr);
 
+public:
 
-        /**
-         * Add a matched writer represented by its attributes.
-         * @param wdata Attributes of the writer to add.
-         * @return True if correctly added.
-         */
-        bool matched_writer_add(const WriterProxyData& wdata) override;
+    /**
+     * Add a matched writer represented by its attributes.
+     * @param wdata Attributes of the writer to add.
+     * @return True if correctly added.
+     */
+    bool matched_writer_add(
+            const WriterProxyData& wdata) override;
 
-        /**
-         * Remove a WriterProxyData from the matached writers.
-         * @param writer_guid GUID of the writer to remove.
-         * @return True if correct.
-         */
-        bool matched_writer_remove(const GUID_t& writer_guid) override;
+    /**
+     * Remove a WriterProxyData from the matached writers.
+     * @param writer_guid GUID of the writer to remove.
+     * @return True if correct.
+     */
+    bool matched_writer_remove(
+            const GUID_t& writer_guid) override;
 
-        /**
-         * Tells us if a specific Writer is matched against this reader.
-         * @param writer_guid GUID of the writer to check.
-         * @return True if it is matched.
-         */
-        bool matched_writer_is_matched(const GUID_t& writer_guid) override;
+    /**
+     * Tells us if a specific Writer is matched against this reader.
+     * @param writer_guid GUID of the writer to check.
+     * @return True if it is matched.
+     */
+    bool matched_writer_is_matched(
+            const GUID_t& writer_guid) override;
 
-        /**
-         * Look for a specific WriterProxy.
-         * @param writerGUID GUID_t of the writer we are looking for.
-         * @param WP Pointer to pointer to a WriterProxy.
-         * @return True if found.
-         */
-        bool matched_writer_lookup(
-                const GUID_t& writerGUID,
-                WriterProxy** WP);
+    /**
+     * Look for a specific WriterProxy.
+     * @param writerGUID GUID_t of the writer we are looking for.
+     * @param WP Pointer to pointer to a WriterProxy.
+     * @return True if found.
+     */
+    bool matched_writer_lookup(
+            const GUID_t& writerGUID,
+            WriterProxy** WP);
 
-        /**
-         * Processes a new DATA message.
-         * @param change Pointer to the CacheChange_t.
-         * @return true if the reader accepts messages.
-         */
-        bool processDataMsg(CacheChange_t* change) override;
+    /**
+     * Processes a new DATA message.
+     * @param change Pointer to the CacheChange_t.
+     * @return true if the reader accepts messages.
+     */
+    bool processDataMsg(
+            CacheChange_t* change) override;
 
-        /**
-         * Processes a new DATA FRAG message.
-         * @param change Pointer to the CacheChange_t.
-         * @param sampleSize Size of the complete assembled message.
-         * @param fragmentStartingNum fragment number of this particular fragment.
-         * @return true if the reader accepts messages.
-         */
-        bool processDataFragMsg(
-                CacheChange_t* change,
-                uint32_t sampleSize,
-                uint32_t fragmentStartingNum) override;
+    /**
+     * Processes a new DATA FRAG message.
+     * @param change Pointer to the CacheChange_t.
+     * @param sampleSize Size of the complete assembled message.
+     * @param fragmentStartingNum fragment number of this particular fragment.
+     * @return true if the reader accepts messages.
+     */
+    bool processDataFragMsg(
+            CacheChange_t* change,
+            uint32_t sampleSize,
+            uint32_t fragmentStartingNum,
+            uint16_t fragmentsInSubmessage) override;
 
-        /**
-         * Processes a new HEARTBEAT message.
-         *
-         * @return true if the reader accepts messages.
-         */
-        bool processHeartbeatMsg(
-                const GUID_t& writerGUID,
-                uint32_t hbCount,
-                const SequenceNumber_t& firstSN,
-                const SequenceNumber_t& lastSN,
-                bool finalFlag,
-                bool livelinessFlag) override;
+    /**
+     * Processes a new HEARTBEAT message.
+     *
+     * @return true if the reader accepts messages.
+     */
+    bool processHeartbeatMsg(
+            const GUID_t& writerGUID,
+            uint32_t hbCount,
+            const SequenceNumber_t& firstSN,
+            const SequenceNumber_t& lastSN,
+            bool finalFlag,
+            bool livelinessFlag) override;
 
-        bool processGapMsg(
-                const GUID_t& writerGUID,
-                const SequenceNumber_t& gapStart,
-                const SequenceNumberSet_t& gapList) override;
+    bool processGapMsg(
+            const GUID_t& writerGUID,
+            const SequenceNumber_t& gapStart,
+            const SequenceNumberSet_t& gapList) override;
 
-        /**
-         * Method to indicate the reader that some change has been removed due to HistoryQos requirements.
-         * @param change Pointer to the CacheChange_t.
-         * @param prox Pointer to the WriterProxy.
-         * @return True if correctly removed.
-         */
-        bool change_removed_by_history(
-                CacheChange_t* change,
-                WriterProxy* prox = nullptr) override;
+    /**
+     * Method to indicate the reader that some change has been removed due to HistoryQos requirements.
+     * @param change Pointer to the CacheChange_t.
+     * @param prox Pointer to the WriterProxy.
+     * @return True if correctly removed.
+     */
+    bool change_removed_by_history(
+            CacheChange_t* change,
+            WriterProxy* prox = nullptr) override;
 
-        /**
-         * This method is called when a new change is received. This method calls the received_change of the History
-         * and depending on the implementation performs different actions.
-         * @param a_change Pointer of the change to add.
-         * @param prox Pointer to the WriterProxy that adds the Change.
-         * @return True if added.
-         */
-        bool change_received(
-                CacheChange_t* a_change,
-                WriterProxy* prox);
+    /**
+     * This method is called when a new change is received. This method calls the received_change of the History
+     * and depending on the implementation performs different actions.
+     * @param a_change Pointer of the change to add.
+     * @param prox Pointer to the WriterProxy that adds the Change.
+     * @return True if added.
+     */
+    bool change_received(
+            CacheChange_t* a_change,
+            WriterProxy* prox);
 
-        /**
-         * Get the RTPS participant
-         * @return Associated RTPS participant
-         */
-        inline RTPSParticipantImpl* getRTPSParticipant() const
-        {
-            return mp_RTPSParticipant;
-        }
+    /**
+     * Get the RTPS participant
+     * @return Associated RTPS participant
+     */
+    inline RTPSParticipantImpl* getRTPSParticipant() const
+    {
+        return mp_RTPSParticipant;
+    }
 
-        /**
-         * Read the next unread CacheChange_t from the history
-         * @param change Pointer to pointer of CacheChange_t
-         * @param wpout Pointer to pointer the matched writer proxy
-         * @return True if read.
-         */
-        bool nextUnreadCache(
-                CacheChange_t** change,
-                WriterProxy** wpout = nullptr) override;
+    /**
+     * Read the next unread CacheChange_t from the history
+     * @param change Pointer to pointer of CacheChange_t
+     * @param wpout Pointer to pointer the matched writer proxy
+     * @return True if read.
+     */
+    bool nextUnreadCache(
+            CacheChange_t** change,
+            WriterProxy** wpout = nullptr) override;
 
-        /**
-         * Take the next CacheChange_t from the history;
-         * @param change Pointer to pointer of CacheChange_t
-         * @param wpout Pointer to pointer the matched writer proxy
-         * @return True if read.
-         */
-        bool nextUntakenCache(
-                CacheChange_t** change,
-                WriterProxy** wpout = nullptr) override;
+    /**
+     * Take the next CacheChange_t from the history;
+     * @param change Pointer to pointer of CacheChange_t
+     * @param wpout Pointer to pointer the matched writer proxy
+     * @return True if read.
+     */
+    bool nextUntakenCache(
+            CacheChange_t** change,
+            WriterProxy** wpout = nullptr) override;
 
-        /**
-         * Update the times parameters of the Reader.
-         * @param times ReaderTimes reference.
-         * @return True if correctly updated.
-         */
-        bool updateTimes(const ReaderTimes& times);
+    /**
+     * Update the times parameters of the Reader.
+     * @param times ReaderTimes reference.
+     * @return True if correctly updated.
+     */
+    bool updateTimes(
+    		const ReaderTimes& times);
 
-        /**
-         *
-         * @return Reference to the ReaderTimes.
-         */
-        inline ReaderTimes& getTimes()
-        {
-            return times_;
-        }
+    /**
+     *
+     * @return Reference to the ReaderTimes.
+     */
+    inline ReaderTimes& getTimes()
+    {
+        return times_;
+    }
 
-        /**
-         * Get the number of matched writers
-         * @return Number of matched writers
-         */
-        inline size_t getMatchedWritersSize() const
-        {
-            return matched_writers_.size();
-        }
+    /**
+     * Get the number of matched writers
+     * @return Number of matched writers
+     */
+    inline size_t getMatchedWritersSize() const
+    {
+        return matched_writers_.size();
+    }
 
-        /*!
-         * @brief Returns there is a clean state with all Writers.
-         * It occurs when the Reader received all samples sent by Writers. In other words,
-         * its WriterProxies are up to date.
-         * @return There is a clean state with all Writers.
-         */
-        bool isInCleanState() override;
+    /*!
+     * @brief Returns there is a clean state with all Writers.
+     * It occurs when the Reader received all samples sent by Writers. In other words,
+     * its WriterProxies are up to date.
+     * @return There is a clean state with all Writers.
+     */
+    bool isInCleanState() override;
 
-        /**
-         * Sends an acknack message from this reader.
-         * @param writer Pointer to the info of the remote writer.
-         * @param sns Sequence number bitmap with the acknack information.
-         * @param sender Message sender interface.
-         * @param is_final Value for final flag.
-         */
-        void send_acknack(
-                const WriterProxy* writer,
-                const SequenceNumberSet_t& sns,
-                const RTPSMessageSenderInterface& sender,
-                bool is_final);
+    /**
+     * Sends an acknack message from this reader.
+     * @param writer Pointer to the info of the remote writer.
+     * @param sns Sequence number bitmap with the acknack information.
+     * @param sender Message sender interface.
+     * @param is_final Value for final flag.
+     */
+    void send_acknack(
+            const WriterProxy* writer,
+            const SequenceNumberSet_t& sns,
+            const RTPSMessageSenderInterface& sender,
+            bool is_final);
 
-        /**
-         * Sends an acknack message from this reader in response to a heartbeat.
-         * @param writer Pointer to the proxy representing the writer to send the acknack to.
-         * @param sender Message sender interface.
-         * @param heartbeat_was_final Final flag of the last received heartbeat.
-         */
-        void send_acknack(
-                const WriterProxy* writer,
-                const RTPSMessageSenderInterface& sender,
-                bool heartbeat_was_final);
+    /**
+     * Sends an acknack message from this reader in response to a heartbeat.
+     * @param writer Pointer to the proxy representing the writer to send the acknack to.
+     * @param sender Message sender interface.
+     * @param heartbeat_was_final Final flag of the last received heartbeat.
+     */
+    void send_acknack(
+            const WriterProxy* writer,
+            const RTPSMessageSenderInterface& sender,
+            bool heartbeat_was_final);
 
-        /**
-         * Use the participant of this reader to send a message to certain locator.
-         * @param message Message to be sent.
-         * @param locator Destination locator.
-         * @param max_blocking_time_point Future time point where any blocking should end.
-         */
-        bool send_sync_nts(
-                CDRMessage_t* message,
-                const Locator_t& locator,
-                std::chrono::steady_clock::time_point& max_blocking_time_point);
+    /**
+     * Use the participant of this reader to send a message to certain locator.
+     * @param message Message to be sent.
+     * @param locator Destination locator.
+     * @param max_blocking_time_point Future time point where any blocking should end.
+     */
+    bool send_sync_nts(
+            CDRMessage_t* message,
+            const Locator_t& locator,
+            std::chrono::steady_clock::time_point& max_blocking_time_point);
 
-    private:
+private:
 
-        bool acceptMsgFrom(
-                const GUID_t& entityGUID,
-                WriterProxy** wp) const;
+    bool acceptMsgFrom(
+            const GUID_t& entityGUID,
+            WriterProxy** wp) const;
 
-        /*!
-         * @remarks Nn thread-safe.
-         */
-        bool findWriterProxy(
-                const GUID_t& writerGUID,
-                WriterProxy** wp) const;
+    /*!
+     * @remarks Nn thread-safe.
+     */
+    bool findWriterProxy(
+            const GUID_t& writerGUID,
+            WriterProxy** wp) const;
 
-        void NotifyChanges(WriterProxy* wp);
+    void NotifyChanges(
+    		WriterProxy* wp);
 
-        //! Acknack Count
-        uint32_t acknack_count_;
-        //! NACKFRAG Count
-        uint32_t nackfrag_count_;
-        //!ReaderTimes of the StatefulReader.
-        ReaderTimes times_;
-        //! Vector containing pointers to all the active WriterProxies.
-        ResourceLimitedVector<WriterProxy*> matched_writers_;
-        //! Vector containing pointers to all the inactive, ready for reuse, WriterProxies.
-        ResourceLimitedVector<WriterProxy*> matched_writers_pool_;
-        //!
-        ResourceLimitedContainerConfig proxy_changes_config_;
-        //! True to disable positive ACKs
-        bool disable_positive_acks_;
-        //! False when being destroyed
-        bool is_alive_;
+    //! Acknack Count
+    uint32_t acknack_count_;
+    //! NACKFRAG Count
+    uint32_t nackfrag_count_;
+    //!ReaderTimes of the StatefulReader.
+    ReaderTimes times_;
+    //! Vector containing pointers to all the active WriterProxies.
+    ResourceLimitedVector<WriterProxy*> matched_writers_;
+    //! Vector containing pointers to all the inactive, ready for reuse, WriterProxies.
+    ResourceLimitedVector<WriterProxy*> matched_writers_pool_;
+    //!
+    ResourceLimitedContainerConfig proxy_changes_config_;
+    //! True to disable positive ACKs
+    bool disable_positive_acks_;
+    //! False when being destroyed
+    bool is_alive_;
 };
 
 } /* namespace rtps */
