+++ include/fastrtps/rtps/reader/StatelessReader.h
@@ -29,18 +29,19 @@
 #include <map>
 
 namespace eprosima {
-namespace fastrtps{
+namespace fastrtps {
 namespace rtps {
 
 /**
  * Class StatelessReader, specialization of the RTPSReader for Best Effort Readers.
  * @ingroup READER_MODULE
  */
-class StatelessReader: public RTPSReader
+class StatelessReader : public RTPSReader
 {
     friend class RTPSParticipantImpl;
 
 public:
+
     virtual ~StatelessReader();
 
 protected:
@@ -59,21 +60,24 @@ public:
      * @param wdata Pointer to the WPD object to add.
      * @return True if correctly added.
      */
-    bool matched_writer_add(const WriterProxyData& wdata) override;
+    bool matched_writer_add(
+            const WriterProxyData& wdata) override;
 
     /**
      * Remove a WriterProxyData from the matached writers.
      * @param writer_guid GUID of the writer to remove.
      * @return True if correct.
      */
-    bool matched_writer_remove(const GUID_t& writer_guid) override;
+    bool matched_writer_remove(
+            const GUID_t& writer_guid) override;
 
     /**
      * Tells us if a specific Writer is matched against this reader.
      * @param writer_guid GUID of the writer to check.
      * @return True if it is matched.
      */
-    bool matched_writer_is_matched(const GUID_t& writer_guid) override;
+    bool matched_writer_is_matched(
+            const GUID_t& writer_guid) override;
 
     /**
      * Method to indicate the reader that some change has been removed due to HistoryQos requirements.
@@ -91,7 +95,8 @@ public:
      * @param change Pointer to the CacheChange_t.
      * @return true if the reader accepts messages from the.
      */
-    bool processDataMsg(CacheChange_t *change) override;
+    bool processDataMsg(
+            CacheChange_t* change) override;
 
     /**
      * Processes a new DATA FRAG message.
@@ -103,7 +108,8 @@ public:
     bool processDataFragMsg(
             CacheChange_t* change,
             uint32_t sampleSize,
-            uint32_t fragmentStartingNum) override;
+            uint32_t fragmentStartingNum,
+            uint16_t fragmentsInSubmessage) override;
 
     /**
      * Processes a new HEARTBEAT message.
@@ -111,7 +117,7 @@ public:
      * @return true if the reader accepts messages from the.
      */
     bool processHeartbeatMsg(
-            const GUID_t &writerGUID,
+            const GUID_t& writerGUID,
             uint32_t hbCount,
             const SequenceNumber_t& firstSN,
             const SequenceNumber_t& lastSN,
@@ -129,7 +135,8 @@ public:
      * @param a_change Pointer of the change to add.
      * @return True if added.
      */
-    bool change_received(CacheChange_t* a_change);
+    bool change_received(
+            CacheChange_t* a_change);
 
     /**
      * Read the next unread CacheChange_t from the history
@@ -155,7 +162,7 @@ public:
      * Get the number of matched writers
      * @return Number of matched writers
      */
-    inline size_t getMatchedWritersSize() const 
+    inline size_t getMatchedWritersSize() const
     {
         return matched_writers_.size();
     }
@@ -166,11 +173,11 @@ public:
      * @return true
      */
     bool isInCleanState() override
-    { 
+    {
         return true;
     }
 
-    inline RTPSParticipantImpl* getRTPSParticipant() const 
+    inline RTPSParticipantImpl* getRTPSParticipant() const
     {
         return mp_RTPSParticipant;
     }
@@ -184,10 +191,11 @@ private:
         bool has_manual_topic_liveliness;
     };
 
-    bool acceptMsgFrom(const GUID_t& entityId);
+    bool acceptMsgFrom(
+            const GUID_t& entityId);
 
     bool thereIsUpperRecordOf(
-            const GUID_t& guid, 
+            const GUID_t& guid,
             const SequenceNumber_t& seq);
 
     /**
@@ -195,7 +203,8 @@ private:
      * @param guid The guid of the remote writer
      * @return True if writer has manual_by_topic livelinesss
      */
-    bool writer_has_manual_liveliness(const GUID_t& guid);
+    bool writer_has_manual_liveliness(
+            const GUID_t& guid);
 
     //!List of GUID_t os matched writers.
     //!Is only used in the Discovery, to correctly notify the user using SubscriptionListener::onSubscriptionMatched();
